Refactor TheMovieDBRepository to load less data
Previously the repository was loading all more data than necessary from
the response from the server since these data were passed directly
between activities once loaded into memory.
Now the repository has a new method to load only the id and poster image
of a movie, wich are the only necessary data for the MainActivity to
work.
This was changes due to the necessity for the DetailsActivity to display
even more data from a movie, so it would become slow if we had to load
at startup all the movies data in one go.